changed folder and file structure
