+++ src/components/Game/CardsGrid/index.js
@@ -1,0 +1,3 @@
+import CardsGrid from './CardsGrid';
+
+export default CardsGrid;
