+++ src/components/Game/LoadingBrain/LoadingBrain.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { ReactComponent as BrainSVG } from '../assets/loading-brain.svg';
-import { LoadingBrainWrapper } from './styles/LoadingBrain.styled';
+import { ReactComponent as BrainSVG } from '../../../assets/loading-brain.svg';
+import { LoadingBrainWrapper } from '../../styles/LoadingBrain.styled';
 
 const LoadingBrain = () => {
     return (
