+++ src/components/Game/LoadingBrain/index.js
@@ -1,0 +1,3 @@
+import LoadingBrain from './LoadingBrain';
+
+export default LoadingBrain;
