+++ src/components/Game/Scores/index.js
@@ -1,0 +1,3 @@
+import Scores from './Scores';
+
+export default Scores;
