+++ src/components/Modal/index.js
@@ -1,0 +1,3 @@
+import Modal from './Modal';
+
+export default Modal;
